requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.8.0
keras>=2.8.0
matplotlib>=3.4.0
streamlit>=1.0.0
Pillow>=8.0.0
joblib>=1.0.0
scikit-learn>=0.24.0
